Added a Vectorizer class and moved CFWrappers there.
